Can now debug with Atmel-ICE or Segger J-Link interchangeably. Notes:
You must now start the gdb server with the current debug<n>s command.

So, for ice, the process is now:

```bash
make debugis
make ${OPTIONS} debugi
# then to kill the server
make debugi-kill
# if the server is stuck
make debugi-murder
```

For the J-Link:


```bash
make debugjs
make ${OPTIONS} debugj
# then to kill the server
make debugj-kill
# if the server is stuck
make debugj-murder
```

As before, the Atmel ICE requires openocd to be installed and in your path.

The JLink requires only the free JLink software to be installed and in your path, NOT openocd as well. (They can both be installed, but it’s not necessary.)